added offline implementation
The commit adds a new OfflineImpl to the API_MANAGER to allow the application to work offline. The get_api_impls Tauri command now returns a HashMap<String, String> containing the name and description of each available API. Additionally, in the App.tsx file, the get_api_impls command is called to retrieve the list of available APIs.